Add bot runner github action, refactor task package to export single-task builder
