+++ Justfile
@@ -63,7 +63,7 @@ test:
 # Run linter
 lint:
     @echo "Running linter..."
-    @which golangci-lint > /dev/null || (echo "Installing golangci-lint..." && go install github.com/golangci/golangci-lint/cmd/golangci-lint@latest)
+    @which golangci-lint > /dev/null || (echo "Installing golangci-lint..." && curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/HEAD/install.sh | sh -s -- -b $(go env GOPATH)/bin v2.3.1)
     golangci-lint run
 
 # Format code
